firmware: implement stat command
Implement the BmcBlobStat command for the generic blob_ids, versus the
active ones.

Also, reduce the size of our transport bits to match the use-case state.

+++ main.cpp
@@ -17,18 +17,18 @@ std::vector<std::string> supportedFirmware = {
 #endif
 };
 
-std::uint32_t supportedTransports =
-    static_cast<std::uint32_t>(FirmwareUpdateFlags::bt);
+std::uint16_t supportedTransports =
+    static_cast<std::uint16_t>(FirmwareUpdateFlags::bt);
 
 void setupFirmwareHandler() __attribute__((constructor));
 
 void setupFirmwareHandler()
 {
 #ifdef ENABLE_PCI_BRIDGE
-    supportedTransports |= static_cast<std::uint32_t>(FirmwareUpdateFlags::p2a);
+    supportedTransports |= static_cast<std::uint16_t>(FirmwareUpdateFlags::p2a);
 #endif
 #ifdef ENABLE_LPC_BRIDGE
-    supportedTransports |= static_cast<std::uint32_t>(FirmwareUpdateFlags::lpc);
+    supportedTransports |= static_cast<std::uint16_t>(FirmwareUpdateFlags::lpc);
 #endif
 
     auto handler = FirmwareBlobHandler::CreateFirmwareBlobHandler(
